+++ config/env.docker.sh
@@ -1,32 +1,35 @@
 ## external
-export API_HOST=http://localhost:8009
-export DEV_TOOLS=false
-export DISCOVERY_HOST=localhost:8000
-export NODE_ENV=production
-export PORT=3000
-export PUBLISH_HOST=hakken
-export WEBPACK_DEVTOOL=cheap-module-eval-source-map
-export SERVICE_NAME=blip
+export API_HOST='http://localhost:8009'
+export DEV_TOOLS='false'
+export SKIP_HAKKEN='false'
+export DISCOVERY_HOST='localhost:8000'
+export NODE_ENV='production'
+export PORT='3000'
+export PUBLISH_HOST='hakken'
+export WEBPACK_DEVTOOL='cheap-module-eval-source-map'
+export SERVICE_NAME='blip'
 
 #Diabeloop specific
-export ALLOW_CHANGE_EMAIL=false
-export I18N_ENABLED=true
-export ALLOW_CHANGE_PASSWORD=false
-export ALLOW_CHANGE_LANG=false
-export ALLOW_SIGNUP_PATIENT=false
-export HIDE_DONATE=true
-export HIDE_DEXCOM_BANNER=true
-export HIDE_UPLOAD_LINK=true
+export ALLOW_CHANGE_EMAIL='false'
+export I18N_ENABLED='true'
+export ALLOW_CHANGE_PASSWORD='false'
+export ALLOW_CHANGE_LANG='false'
+export ALLOW_SIGNUP_PATIENT='false'
+export HIDE_DONATE='true'
+export HIDE_DEXCOM_BANNER='true'
+export HIDE_UPLOAD_LINK='true'
 export BRANDING='diabeloop'
-export PASSWORD_MIN_LENGTH=10
-export PASSWORD_MAX_LENGTH=72
+export PASSWORD_MIN_LENGTH='10'
+export PASSWORD_MAX_LENGTH='72'
 # url to HELP system
-export HELP_LINK='disable'
+export HELP_LINK='disabled'
 # url to web server hosting the asset files
 # such as terms of use, data privacy, etc.
 export ASSETS_URL='https://url.com'
-# Matomo tracking URL (use disable to desactivate tracking)
-export MATOMO_TRACKER_URL='disable'
+# Metrics service to use (disabled, highwater, matomo)
+export METRICS_SERVICE='disabled'
+# Metrics service Matomo URL (used only when METRICS_SERVICE='matomo'):
+export MATOMO_TRACKER_URL='disabled'
 export MAX_FAILED_LOGIN_ATTEMPTS='5'
 export DELAY_BEFORE_NEXT_LOGIN_ATTEMPT='1'
 export TERMS_PRIVACY_DATE='2020-02-26'