+++ config/local.sh
@@ -1,18 +1,23 @@
 export API_HOST='http://localhost:8009'
-export I18N_ENABLED=false
-export ALLOW_SIGNUP_PATIENT=true
-export ALLOW_PATIENT_CHANGE_EMAIL=true
-export ALLOW_PATIENT_CHANGE_PASSWORD=true
-export CAN_SEE_PWD_LOGIN=true
+export I18N_ENABLED='false'
+export SKIP_HAKKEN='true'
+export ALLOW_SIGNUP_PATIENT='true'
+export ALLOW_PATIENT_CHANGE_EMAIL='true'
+export ALLOW_PATIENT_CHANGE_PASSWORD='true'
+export CAN_SEE_PWD_LOGIN='true'
 export SUPPORT_EMAIL_ADDRESS='yourloops@diabeloop.com'
 export SUPPORT_WEB_ADDRESS='https://support.diabeloop.com/hc/categories/360001386093-YourLoops'
-export HIDE_UPLOAD_LINK=false
-export HIDE_DEXCOM_BANNER=false
-export HIDE_DONATE=false
+export HIDE_UPLOAD_LINK='false'
+export HIDE_DEXCOM_BANNER='false'
+export HIDE_DONATE='false'
 export BRANDING='tidepool'
 export SERVICE_NAME='blip'
-# Matomo tracking URL (use disable to desactivate tracking)
-export MATOMO_TRACKER_URL='disable'
+# url to HELP system
+export HELP_LINK='disabled'
+# Metrics service to use ('disabled', 'highwater', 'matomo')
+export METRICS_SERVICE='disabled'
+# Metrics service Matomo URL (used only when METRICS_SERVICE='matomo'):
+export MATOMO_TRACKER_URL='disabled'
 export MAX_FAILED_LOGIN_ATTEMPTS='5'
 export DELAY_BEFORE_NEXT_LOGIN_ATTEMPT='10'
 export TERMS_PRIVACY_DATE='2020-02-26'