+++ CHANGELOG.md
@@ -5,6 +5,7 @@ It is based on Tidepool Blip 1.27.
 ## Unreleased
 ### Changed
 - PT-1205 Add timezone info on tooltips when necessary
+- PT-1254 Disable Highwater from Blip
 - PT-1256 Improve PDF generation
 ### Engineering use
 - PT-1249 Reduce blip & viz build time.